+++ keanu-project/src/main/java/io/improbable/keanu/vertices/bool/nonprobabilistic/BooleanIfVertex.java
@@ -12,7 +12,7 @@ public class BooleanIfVertex extends BoolVertex implements NonProbabilistic<Bool
     private final Vertex<? extends BooleanTensor> thn;
     private final Vertex<? extends BooleanTensor> els;
 
-    public BooleanIfVertex(int[] shape,
+    public BooleanIfVertex(long[] shape,
                            Vertex<? extends BooleanTensor> predicate,
                            Vertex<? extends BooleanTensor> thn,
                            Vertex<? extends BooleanTensor> els) {
